Struct rusoto_devicefarm::CreateNetworkProfileRequest

pub struct CreateNetworkProfileRequest {
    pub description: Option<String>,
    pub downlink_bandwidth_bits: Option<i64>,
    pub downlink_delay_ms: Option<i64>,
    pub downlink_jitter_ms: Option<i64>,
    pub downlink_loss_percent: Option<i64>,
    pub name: String,
    pub project_arn: String,
    pub type_: Option<String>,
    pub uplink_bandwidth_bits: Option<i64>,
    pub uplink_delay_ms: Option<i64>,
    pub uplink_jitter_ms: Option<i64>,
    pub uplink_loss_percent: Option<i64>,
}

Fields
description: Option<String>

The description of the network profile.

downlink_bandwidth_bits: Option<i64>

The data throughput rate in bits per second, as an integer from 0 to 104857600.

downlink_delay_ms: Option<i64>

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

downlink_jitter_ms: Option<i64>

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

downlink_loss_percent: Option<i64>

Proportion of received packets that fail to arrive from 0 to 100 percent.

name: String

The name you wish to specify for the new network profile.

project_arn: String

The Amazon Resource Name (ARN) of the project for which you want to create a network profile.

type_: Option<String>

The type of network profile you wish to create. Valid values are listed below.

uplink_bandwidth_bits: Option<i64>

The data throughput rate in bits per second, as an integer from 0 to 104857600.

uplink_delay_ms: Option<i64>

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

uplink_jitter_ms: Option<i64>

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

uplink_loss_percent: Option<i64>

Proportion of transmitted packets that fail to arrive from 0 to 100 percent.