Struct rusoto_devicefarm::Device

pub struct Device {
    pub arn: Option<String>,
    pub carrier: Option<String>,
    pub cpu: Option<CPU>,
    pub fleet_name: Option<String>,
    pub fleet_type: Option<String>,
    pub form_factor: Option<String>,
    pub heap_size: Option<i64>,
    pub image: Option<String>,
    pub manufacturer: Option<String>,
    pub memory: Option<i64>,
    pub model: Option<String>,
    pub name: Option<String>,
    pub os: Option<String>,
    pub platform: Option<String>,
    pub radio: Option<String>,
    pub remote_access_enabled: Option<bool>,
    pub resolution: Option<Resolution>,
}

Represents a device type that an app is tested against.

Fields
arn: Option<String>

The device's ARN.

carrier: Option<String>

The device's carrier.

cpu: Option<CPU>

Information about the device's CPU.

fleet_name: Option<String>

The name of the fleet to which this device belongs.

fleet_type: Option<String>

The type of fleet to which this device belongs. Possible values for fleet type are PRIVATE and PUBLIC.

form_factor: Option<String>

The device's form factor.

Allowed values include:

PHONE: The phone form factor.

TABLET: The tablet form factor.

heap_size: Option<i64>

The device's heap size, expressed in bytes.

image: Option<String>

The device's image name.

manufacturer: Option<String>

The device's manufacturer name.

memory: Option<i64>

The device's total memory size, expressed in bytes.

model: Option<String>

The device's model name.

name: Option<String>

The device's display name.

os: Option<String>

The device's operating system type.

platform: Option<String>

The device's platform.

Allowed values include:

ANDROID: The Android platform.

IOS: The iOS platform.

radio: Option<String>

The device's radio.

remote_access_enabled: Option<bool>

Specifies whether remote access has been enabled for the specified device.

resolution: Option<Resolution>

The resolution of the device.