Struct rusoto_devicefarm::RemoteAccessSession

pub struct RemoteAccessSession {
    pub arn: Option<String>,
    pub billing_method: Option<String>,
    pub client_id: Option<String>,
    pub created: Option<f64>,
    pub device: Option<Device>,
    pub device_minutes: Option<DeviceMinutes>,
    pub device_udid: Option<String>,
    pub endpoint: Option<String>,
    pub host_address: Option<String>,
    pub instance_arn: Option<String>,
    pub interaction_mode: Option<String>,
    pub message: Option<String>,
    pub name: Option<String>,
    pub remote_debug_enabled: Option<bool>,
    pub remote_record_app_arn: Option<String>,
    pub remote_record_enabled: Option<bool>,
    pub result: Option<String>,
    pub skip_app_resign: Option<bool>,
    pub started: Option<f64>,
    pub status: Option<String>,
    pub stopped: Option<f64>,
}

Represents information about the remote access session.

Fields
arn: Option<String>

The Amazon Resource Name (ARN) of the remote access session.

billing_method: Option<String>

The billing method of the remote access session. Possible values include METERED or UNMETERED. For more information about metered devices, see AWS Device Farm terminology."

client_id: Option<String>

Unique identifier of your client for the remote access session. Only returned if remote debugging is enabled for the remote access session.

created: Option<f64>

The date and time the remote access session was created.

device: Option<Device>

The device (phone or tablet) used in the remote access session.

device_minutes: Option<DeviceMinutes>

The number of minutes a device is used in a remote access sesssion (including setup and teardown minutes).

device_udid: Option<String>

Unique device identifier for the remote device. Only returned if remote debugging is enabled for the remote access session.

endpoint: Option<String>

The endpoint for the remote access sesssion.

host_address: Option<String>

IP address of the EC2 host where you need to connect to remotely debug devices. Only returned if remote debugging is enabled for the remote access session.

instance_arn: Option<String>

The Amazon Resource Name (ARN) of the instance.

interaction_mode: Option<String>

The interaction mode of the remote access session. Valid values are:

INTERACTIVE: You can interact with the iOS device by viewing, touching, and rotating the screen. You cannot run XCUITest framework-based tests in this mode.

NOVIDEO: You are connected to the device but cannot interact with it or view the screen. This mode has the fastest test execution speed. You can run XCUITest framework-based tests in this mode.

VIDEOONLY: You can view the screen but cannot touch or rotate it. You can run XCUITest framework-based tests and watch the screen in this mode.

message: Option<String>

A message about the remote access session.

name: Option<String>

The name of the remote access session.

remote_debug_enabled: Option<bool>

This flag is set to true if remote debugging is enabled for the remote access session.

remote_record_app_arn: Option<String>

The Amazon Resource Name (ARN) for the app to be recorded in the remote access session.

remote_record_enabled: Option<bool>

This flag is set to true if remote recording is enabled for the remote access session.

result: Option<String>

The result of the remote access session. Can be any of the following:

PENDING: A pending condition.

PASSED: A passing condition.

WARNED: A warning condition.

FAILED: A failed condition.

SKIPPED: A skipped condition.

ERRORED: An error condition.

STOPPED: A stopped condition.

skip_app_resign: Option<bool>

When set to true, for private devices, Device Farm will not sign your app again. For public devices, Device Farm always signs your apps again and this parameter has no effect.

For more information about how Device Farm re-signs your app(s), see Do you modify my app? in the AWS Device Farm FAQs.

started: Option<f64>

The date and time the remote access session was started.

status: Option<String>

The status of the remote access session. Can be any of the following:

PENDING: A pending status.

PENDINGCONCURRENCY: A pending concurrency status.

PENDINGDEVICE: A pending device status.

PROCESSING: A processing status.

SCHEDULING: A scheduling status.

PREPARING: A preparing status.

RUNNING: A running status.

COMPLETED: A completed status.

STOPPING: A stopping status.

stopped: Option<f64>

The date and time the remote access session was stopped.