Struct rusoto_devicefarm::ScheduleRunTest

pub struct ScheduleRunTest {
    pub filter: Option<String>,
    pub parameters: Option<HashMap<String, String>>,
    pub test_package_arn: Option<String>,
    pub type_: String,
}

Represents additional test settings.

Fields
filter: Option<String>

The test's filter.

parameters: Option<HashMap<String, String>>

The test's parameters, such as the following test framework parameters and fixture settings:

For Calabash tests:

profile: A cucumber profile, for example, "my_profile_name".

tags: You can limit execution to features or scenarios that have (or don't have) certain tags, for example, "@smoke" or "@smoke,~@wip".

For Appium tests (all types):

appium_version: The Appium version. Currently supported values are "1.4.16", "1.6.3", "latest", and "default".

“latest” will run the latest Appium version supported by Device Farm (1.6.3).

For “default”, Device Farm will choose a compatible version of Appium for the device. The current behavior is to run 1.4.16 on Android devices and iOS 9 and earlier, 1.6.3 for iOS 10 and later.

This behavior is subject to change.

For Fuzz tests (Android only):

event_count: The number of events, between 1 and 10000, that the UI fuzz test should perform.

throttle: The time, in ms, between 0 and 1000, that the UI fuzz test should wait between events.

seed: A seed to use for randomizing the UI fuzz test. Using the same seed value between tests ensures identical event sequences.

For Explorer tests:

username: A username to use if the Explorer encounters a login form. If not supplied, no username will be inserted.

password: A password to use if the Explorer encounters a login form. If not supplied, no password will be inserted.

For Instrumentation:

filter: A test filter string. Examples:

Running a single test case: "com.android.abc.Test1"

Running a single test: "com.android.abc.Test1#smoke"

Running multiple tests: "com.android.abc.Test1,com.android.abc.Test2"

For XCTest and XCTestUI:

filter: A test filter string. Examples:

Running a single test class: "LoginTests"

Running a multiple test classes: "LoginTests,SmokeTests"

Running a single test: "LoginTests/testValid"

Running multiple tests: "LoginTests/testValid,LoginTests/testInvalid"

For UIAutomator:

filter: A test filter string. Examples:

Running a single test case: "com.android.abc.Test1"

Running a single test: "com.android.abc.Test1#smoke"

Running multiple tests: "com.android.abc.Test1,com.android.abc.Test2"

test_package_arn: Option<String>

The ARN of the uploaded test that will be run.

type_: String

The test's type.

Must be one of the following values:

BUILTIN_FUZZ: The built-in fuzz type.

BUILTIN_EXPLORER: For Android, an app explorer that will traverse an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT: The Appium Java JUnit type.

APPIUM_JAVA_TESTNG: The Appium Java TestNG type.

APPIUM_PYTHON: The Appium Python type.

APPIUM_WEB_JAVA_JUNIT: The Appium Java JUnit type for Web apps.

APPIUM_WEB_JAVA_TESTNG: The Appium Java TestNG type for Web apps.

APPIUM_WEB_PYTHON: The Appium Python type for Web apps.

CALABASH: The Calabash type.

INSTRUMENTATION: The Instrumentation type.

UIAUTOMATION: The uiautomation type.

UIAUTOMATOR: The uiautomator type.

XCTEST: The XCode test type.

XCTEST_UI: The XCode UI test type.