Trait rusoto_devicefarm::DeviceFarm

pub trait DeviceFarm {
    fn create_device_pool(
        &self, 
        input: CreateDevicePoolRequest
    ) -> RusotoFuture<CreateDevicePoolResult, CreateDevicePoolError>;

    fn create_instance_profile(
        &self, 
        input: CreateInstanceProfileRequest
    ) -> RusotoFuture<CreateInstanceProfileResult, CreateInstanceProfileError>;

    fn create_network_profile(
        &self, 
        input: CreateNetworkProfileRequest
    ) -> RusotoFuture<CreateNetworkProfileResult, CreateNetworkProfileError>;

    fn create_project(
        &self, 
        input: CreateProjectRequest
    ) -> RusotoFuture<CreateProjectResult, CreateProjectError>;

    fn create_remote_access_session(
        &self, 
        input: CreateRemoteAccessSessionRequest
    ) -> RusotoFuture<CreateRemoteAccessSessionResult, CreateRemoteAccessSessionError>;

    fn create_upload(
        &self, 
        input: CreateUploadRequest
    ) -> RusotoFuture<CreateUploadResult, CreateUploadError>;

    fn create_vpce_configuration(
        &self, 
        input: CreateVPCEConfigurationRequest
    ) -> RusotoFuture<CreateVPCEConfigurationResult, CreateVPCEConfigurationError>;

    fn delete_device_pool(
        &self, 
        input: DeleteDevicePoolRequest
    ) -> RusotoFuture<DeleteDevicePoolResult, DeleteDevicePoolError>;

    fn delete_instance_profile(
        &self, 
        input: DeleteInstanceProfileRequest
    ) -> RusotoFuture<DeleteInstanceProfileResult, DeleteInstanceProfileError>;

    fn delete_network_profile(
        &self, 
        input: DeleteNetworkProfileRequest
    ) -> RusotoFuture<DeleteNetworkProfileResult, DeleteNetworkProfileError>;

    fn delete_project(
        &self, 
        input: DeleteProjectRequest
    ) -> RusotoFuture<DeleteProjectResult, DeleteProjectError>;

    fn delete_remote_access_session(
        &self, 
        input: DeleteRemoteAccessSessionRequest
    ) -> RusotoFuture<DeleteRemoteAccessSessionResult, DeleteRemoteAccessSessionError>;

    fn delete_run(
        &self, 
        input: DeleteRunRequest
    ) -> RusotoFuture<DeleteRunResult, DeleteRunError>;

    fn delete_upload(
        &self, 
        input: DeleteUploadRequest
    ) -> RusotoFuture<DeleteUploadResult, DeleteUploadError>;

    fn delete_vpce_configuration(
        &self, 
        input: DeleteVPCEConfigurationRequest
    ) -> RusotoFuture<DeleteVPCEConfigurationResult, DeleteVPCEConfigurationError>;

    fn get_account_settings(
        &self
    ) -> RusotoFuture<GetAccountSettingsResult, GetAccountSettingsError>;

    fn get_device(
        &self, 
        input: GetDeviceRequest
    ) -> RusotoFuture<GetDeviceResult, GetDeviceError>;

    fn get_device_instance(
        &self, 
        input: GetDeviceInstanceRequest
    ) -> RusotoFuture<GetDeviceInstanceResult, GetDeviceInstanceError>;

    fn get_device_pool(
        &self, 
        input: GetDevicePoolRequest
    ) -> RusotoFuture<GetDevicePoolResult, GetDevicePoolError>;

    fn get_device_pool_compatibility(
        &self, 
        input: GetDevicePoolCompatibilityRequest
    ) -> RusotoFuture<GetDevicePoolCompatibilityResult, GetDevicePoolCompatibilityError>;

    fn get_instance_profile(
        &self, 
        input: GetInstanceProfileRequest
    ) -> RusotoFuture<GetInstanceProfileResult, GetInstanceProfileError>;

    fn get_job(
        &self, 
        input: GetJobRequest
    ) -> RusotoFuture<GetJobResult, GetJobError>;

    fn get_network_profile(
        &self, 
        input: GetNetworkProfileRequest
    ) -> RusotoFuture<GetNetworkProfileResult, GetNetworkProfileError>;

    fn get_offering_status(
        &self, 
        input: GetOfferingStatusRequest
    ) -> RusotoFuture<GetOfferingStatusResult, GetOfferingStatusError>;

    fn get_project(
        &self, 
        input: GetProjectRequest
    ) -> RusotoFuture<GetProjectResult, GetProjectError>;

    fn get_remote_access_session(
        &self, 
        input: GetRemoteAccessSessionRequest
    ) -> RusotoFuture<GetRemoteAccessSessionResult, GetRemoteAccessSessionError>;

    fn get_run(
        &self, 
        input: GetRunRequest
    ) -> RusotoFuture<GetRunResult, GetRunError>;

    fn get_suite(
        &self, 
        input: GetSuiteRequest
    ) -> RusotoFuture<GetSuiteResult, GetSuiteError>;

    fn get_test(
        &self, 
        input: GetTestRequest
    ) -> RusotoFuture<GetTestResult, GetTestError>;

    fn get_upload(
        &self, 
        input: GetUploadRequest
    ) -> RusotoFuture<GetUploadResult, GetUploadError>;

    fn get_vpce_configuration(
        &self, 
        input: GetVPCEConfigurationRequest
    ) -> RusotoFuture<GetVPCEConfigurationResult, GetVPCEConfigurationError>;

    fn install_to_remote_access_session(
        &self, 
        input: InstallToRemoteAccessSessionRequest
    ) -> RusotoFuture<InstallToRemoteAccessSessionResult, InstallToRemoteAccessSessionError>;

    fn list_artifacts(
        &self, 
        input: ListArtifactsRequest
    ) -> RusotoFuture<ListArtifactsResult, ListArtifactsError>;

    fn list_device_instances(
        &self, 
        input: ListDeviceInstancesRequest
    ) -> RusotoFuture<ListDeviceInstancesResult, ListDeviceInstancesError>;

    fn list_device_pools(
        &self, 
        input: ListDevicePoolsRequest
    ) -> RusotoFuture<ListDevicePoolsResult, ListDevicePoolsError>;

    fn list_devices(
        &self, 
        input: ListDevicesRequest
    ) -> RusotoFuture<ListDevicesResult, ListDevicesError>;

    fn list_instance_profiles(
        &self, 
        input: ListInstanceProfilesRequest
    ) -> RusotoFuture<ListInstanceProfilesResult, ListInstanceProfilesError>;

    fn list_jobs(
        &self, 
        input: ListJobsRequest
    ) -> RusotoFuture<ListJobsResult, ListJobsError>;

    fn list_network_profiles(
        &self, 
        input: ListNetworkProfilesRequest
    ) -> RusotoFuture<ListNetworkProfilesResult, ListNetworkProfilesError>;

    fn list_offering_promotions(
        &self, 
        input: ListOfferingPromotionsRequest
    ) -> RusotoFuture<ListOfferingPromotionsResult, ListOfferingPromotionsError>;

    fn list_offering_transactions(
        &self, 
        input: ListOfferingTransactionsRequest
    ) -> RusotoFuture<ListOfferingTransactionsResult, ListOfferingTransactionsError>;

    fn list_offerings(
        &self, 
        input: ListOfferingsRequest
    ) -> RusotoFuture<ListOfferingsResult, ListOfferingsError>;

    fn list_projects(
        &self, 
        input: ListProjectsRequest
    ) -> RusotoFuture<ListProjectsResult, ListProjectsError>;

    fn list_remote_access_sessions(
        &self, 
        input: ListRemoteAccessSessionsRequest
    ) -> RusotoFuture<ListRemoteAccessSessionsResult, ListRemoteAccessSessionsError>;

    fn list_runs(
        &self, 
        input: ListRunsRequest
    ) -> RusotoFuture<ListRunsResult, ListRunsError>;

    fn list_samples(
        &self, 
        input: ListSamplesRequest
    ) -> RusotoFuture<ListSamplesResult, ListSamplesError>;

    fn list_suites(
        &self, 
        input: ListSuitesRequest
    ) -> RusotoFuture<ListSuitesResult, ListSuitesError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn list_tests(
        &self, 
        input: ListTestsRequest
    ) -> RusotoFuture<ListTestsResult, ListTestsError>;

    fn list_unique_problems(
        &self, 
        input: ListUniqueProblemsRequest
    ) -> RusotoFuture<ListUniqueProblemsResult, ListUniqueProblemsError>;

    fn list_uploads(
        &self, 
        input: ListUploadsRequest
    ) -> RusotoFuture<ListUploadsResult, ListUploadsError>;

    fn list_vpce_configurations(
        &self, 
        input: ListVPCEConfigurationsRequest
    ) -> RusotoFuture<ListVPCEConfigurationsResult, ListVPCEConfigurationsError>;

    fn purchase_offering(
        &self, 
        input: PurchaseOfferingRequest
    ) -> RusotoFuture<PurchaseOfferingResult, PurchaseOfferingError>;

    fn renew_offering(
        &self, 
        input: RenewOfferingRequest
    ) -> RusotoFuture<RenewOfferingResult, RenewOfferingError>;

    fn schedule_run(
        &self, 
        input: ScheduleRunRequest
    ) -> RusotoFuture<ScheduleRunResult, ScheduleRunError>;

    fn stop_job(
        &self, 
        input: StopJobRequest
    ) -> RusotoFuture<StopJobResult, StopJobError>;

    fn stop_remote_access_session(
        &self, 
        input: StopRemoteAccessSessionRequest
    ) -> RusotoFuture<StopRemoteAccessSessionResult, StopRemoteAccessSessionError>;

    fn stop_run(
        &self, 
        input: StopRunRequest
    ) -> RusotoFuture<StopRunResult, StopRunError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn update_device_instance(
        &self, 
        input: UpdateDeviceInstanceRequest
    ) -> RusotoFuture<UpdateDeviceInstanceResult, UpdateDeviceInstanceError>;

    fn update_device_pool(
        &self, 
        input: UpdateDevicePoolRequest
    ) -> RusotoFuture<UpdateDevicePoolResult, UpdateDevicePoolError>;

    fn update_instance_profile(
        &self, 
        input: UpdateInstanceProfileRequest
    ) -> RusotoFuture<UpdateInstanceProfileResult, UpdateInstanceProfileError>;

    fn update_network_profile(
        &self, 
        input: UpdateNetworkProfileRequest
    ) -> RusotoFuture<UpdateNetworkProfileResult, UpdateNetworkProfileError>;

    fn update_project(
        &self, 
        input: UpdateProjectRequest
    ) -> RusotoFuture<UpdateProjectResult, UpdateProjectError>;

    fn update_upload(
        &self, 
        input: UpdateUploadRequest
    ) -> RusotoFuture<UpdateUploadResult, UpdateUploadError>;

    fn update_vpce_configuration(
        &self, 
        input: UpdateVPCEConfigurationRequest
    ) -> RusotoFuture<UpdateVPCEConfigurationResult, UpdateVPCEConfigurationError>;
}

Trait representing the capabilities of the AWS Device Farm API. AWS Device Farm clients implement this trait.

Required methods

fn create_device_pool(
&self,
input: CreateDevicePoolRequest
) -> RusotoFuture<CreateDevicePoolResult, CreateDevicePoolError>

Creates a device pool.

fn create_instance_profile(
&self,
input: CreateInstanceProfileRequest
) -> RusotoFuture<CreateInstanceProfileResult, CreateInstanceProfileError>

Creates a profile that can be applied to one or more private fleet device instances.

fn create_network_profile(
&self,
input: CreateNetworkProfileRequest
) -> RusotoFuture<CreateNetworkProfileResult, CreateNetworkProfileError>

Creates a network profile.

fn create_project(
&self,
input: CreateProjectRequest
) -> RusotoFuture<CreateProjectResult, CreateProjectError>

Creates a new project.

fn create_remote_access_session(
&self,
input: CreateRemoteAccessSessionRequest
) -> RusotoFuture<CreateRemoteAccessSessionResult, CreateRemoteAccessSessionError>

Specifies and starts a remote access session.

fn create_upload(
&self,
input: CreateUploadRequest
) -> RusotoFuture<CreateUploadResult, CreateUploadError>

Uploads an app or test scripts.

fn create_vpce_configuration(
&self,
input: CreateVPCEConfigurationRequest
) -> RusotoFuture<CreateVPCEConfigurationResult, CreateVPCEConfigurationError>

Creates a configuration record in Device Farm for your Amazon Virtual Private Cloud (VPC) endpoint.

fn delete_device_pool(
&self,
input: DeleteDevicePoolRequest
) -> RusotoFuture<DeleteDevicePoolResult, DeleteDevicePoolError>

Deletes a device pool given the pool ARN. Does not allow deletion of curated pools owned by the system.

fn delete_instance_profile(
&self,
input: DeleteInstanceProfileRequest
) -> RusotoFuture<DeleteInstanceProfileResult, DeleteInstanceProfileError>

Deletes a profile that can be applied to one or more private device instances.

fn delete_network_profile(
&self,
input: DeleteNetworkProfileRequest
) -> RusotoFuture<DeleteNetworkProfileResult, DeleteNetworkProfileError>

Deletes a network profile.

fn delete_project(
&self,
input: DeleteProjectRequest
) -> RusotoFuture<DeleteProjectResult, DeleteProjectError>

Deletes an AWS Device Farm project, given the project ARN.

Note Deleting this resource does not stop an in-progress run.

fn delete_remote_access_session(
&self,
input: DeleteRemoteAccessSessionRequest
) -> RusotoFuture<DeleteRemoteAccessSessionResult, DeleteRemoteAccessSessionError>

Deletes a completed remote access session and its results.

fn delete_run(
&self,
input: DeleteRunRequest
) -> RusotoFuture<DeleteRunResult, DeleteRunError>

Deletes the run, given the run ARN.

Note Deleting this resource does not stop an in-progress run.

fn delete_upload(
&self,
input: DeleteUploadRequest
) -> RusotoFuture<DeleteUploadResult, DeleteUploadError>

Deletes an upload given the upload ARN.

fn delete_vpce_configuration(
&self,
input: DeleteVPCEConfigurationRequest
) -> RusotoFuture<DeleteVPCEConfigurationResult, DeleteVPCEConfigurationError>

Deletes a configuration for your Amazon Virtual Private Cloud (VPC) endpoint.

fn get_account_settings(
&self
) -> RusotoFuture<GetAccountSettingsResult, GetAccountSettingsError>

Returns the number of unmetered iOS and/or unmetered Android devices that have been purchased by the account.

fn get_device(
&self,
input: GetDeviceRequest
) -> RusotoFuture<GetDeviceResult, GetDeviceError>

Gets information about a unique device type.

fn get_device_instance(
&self,
input: GetDeviceInstanceRequest
) -> RusotoFuture<GetDeviceInstanceResult, GetDeviceInstanceError>

Returns information about a device instance belonging to a private device fleet.

fn get_device_pool(
&self,
input: GetDevicePoolRequest
) -> RusotoFuture<GetDevicePoolResult, GetDevicePoolError>

Gets information about a device pool.

fn get_device_pool_compatibility(
&self,
input: GetDevicePoolCompatibilityRequest
) -> RusotoFuture<GetDevicePoolCompatibilityResult, GetDevicePoolCompatibilityError>

Gets information about compatibility with a device pool.

fn get_instance_profile(
&self,
input: GetInstanceProfileRequest
) -> RusotoFuture<GetInstanceProfileResult, GetInstanceProfileError>

Returns information about the specified instance profile.

fn get_job(
&self,
input: GetJobRequest
) -> RusotoFuture<GetJobResult, GetJobError>

Gets information about a job.

fn get_network_profile(
&self,
input: GetNetworkProfileRequest
) -> RusotoFuture<GetNetworkProfileResult, GetNetworkProfileError>

Returns information about a network profile.

fn get_offering_status(
&self,
input: GetOfferingStatusRequest
) -> RusotoFuture<GetOfferingStatusResult, GetOfferingStatusError>

Gets the current status and future status of all offerings purchased by an AWS account. The response indicates how many offerings are currently available and the offerings that will be available in the next period. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

fn get_project(
&self,
input: GetProjectRequest
) -> RusotoFuture<GetProjectResult, GetProjectError>

Gets information about a project.

fn get_remote_access_session(
&self,
input: GetRemoteAccessSessionRequest
) -> RusotoFuture<GetRemoteAccessSessionResult, GetRemoteAccessSessionError>

Returns a link to a currently running remote access session.

fn get_run(
&self,
input: GetRunRequest
) -> RusotoFuture<GetRunResult, GetRunError>

Gets information about a run.

fn get_suite(
&self,
input: GetSuiteRequest
) -> RusotoFuture<GetSuiteResult, GetSuiteError>

Gets information about a suite.

fn get_test(
&self,
input: GetTestRequest
) -> RusotoFuture<GetTestResult, GetTestError>

Gets information about a test.

fn get_upload(
&self,
input: GetUploadRequest
) -> RusotoFuture<GetUploadResult, GetUploadError>

Gets information about an upload.

fn get_vpce_configuration(
&self,
input: GetVPCEConfigurationRequest
) -> RusotoFuture<GetVPCEConfigurationResult, GetVPCEConfigurationError>

Returns information about the configuration settings for your Amazon Virtual Private Cloud (VPC) endpoint.

fn install_to_remote_access_session(
&self,
input: InstallToRemoteAccessSessionRequest
) -> RusotoFuture<InstallToRemoteAccessSessionResult, InstallToRemoteAccessSessionError>

Installs an application to the device in a remote access session. For Android applications, the file must be in .apk format. For iOS applications, the file must be in .ipa format.

fn list_artifacts(
&self,
input: ListArtifactsRequest
) -> RusotoFuture<ListArtifactsResult, ListArtifactsError>

Gets information about artifacts.

fn list_device_instances(
&self,
input: ListDeviceInstancesRequest
) -> RusotoFuture<ListDeviceInstancesResult, ListDeviceInstancesError>

Returns information about the private device instances associated with one or more AWS accounts.

fn list_device_pools(
&self,
input: ListDevicePoolsRequest
) -> RusotoFuture<ListDevicePoolsResult, ListDevicePoolsError>

Gets information about device pools.

fn list_devices(
&self,
input: ListDevicesRequest
) -> RusotoFuture<ListDevicesResult, ListDevicesError>

Gets information about unique device types.

fn list_instance_profiles(
&self,
input: ListInstanceProfilesRequest
) -> RusotoFuture<ListInstanceProfilesResult, ListInstanceProfilesError>

Returns information about all the instance profiles in an AWS account.

fn list_jobs(
&self,
input: ListJobsRequest
) -> RusotoFuture<ListJobsResult, ListJobsError>

Gets information about jobs for a given test run.

fn list_network_profiles(
&self,
input: ListNetworkProfilesRequest
) -> RusotoFuture<ListNetworkProfilesResult, ListNetworkProfilesError>

Returns the list of available network profiles.

fn list_offering_promotions(
&self,
input: ListOfferingPromotionsRequest
) -> RusotoFuture<ListOfferingPromotionsResult, ListOfferingPromotionsError>

Returns a list of offering promotions. Each offering promotion record contains the ID and description of the promotion. The API returns a NotEligible error if the caller is not permitted to invoke the operation. Contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

fn list_offering_transactions(
&self,
input: ListOfferingTransactionsRequest
) -> RusotoFuture<ListOfferingTransactionsResult, ListOfferingTransactionsError>

Returns a list of all historical purchases, renewals, and system renewal transactions for an AWS account. The list is paginated and ordered by a descending timestamp (most recent transactions are first). The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

fn list_offerings(
&self,
input: ListOfferingsRequest
) -> RusotoFuture<ListOfferingsResult, ListOfferingsError>

Returns a list of products or offerings that the user can manage through the API. Each offering record indicates the recurring price per unit and the frequency for that offering. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

fn list_projects(
&self,
input: ListProjectsRequest
) -> RusotoFuture<ListProjectsResult, ListProjectsError>

Gets information about projects.

fn list_remote_access_sessions(
&self,
input: ListRemoteAccessSessionsRequest
) -> RusotoFuture<ListRemoteAccessSessionsResult, ListRemoteAccessSessionsError>

Returns a list of all currently running remote access sessions.

fn list_runs(
&self,
input: ListRunsRequest
) -> RusotoFuture<ListRunsResult, ListRunsError>

Gets information about runs, given an AWS Device Farm project ARN.

fn list_samples(
&self,
input: ListSamplesRequest
) -> RusotoFuture<ListSamplesResult, ListSamplesError>

Gets information about samples, given an AWS Device Farm job ARN.

fn list_suites(
&self,
input: ListSuitesRequest
) -> RusotoFuture<ListSuitesResult, ListSuitesError>

Gets information about test suites for a given job.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

List the tags for an AWS Device Farm resource.

fn list_tests(
&self,
input: ListTestsRequest
) -> RusotoFuture<ListTestsResult, ListTestsError>

Gets information about tests in a given test suite.

fn list_unique_problems(
&self,
input: ListUniqueProblemsRequest
) -> RusotoFuture<ListUniqueProblemsResult, ListUniqueProblemsError>

Gets information about unique problems.

fn list_uploads(
&self,
input: ListUploadsRequest
) -> RusotoFuture<ListUploadsResult, ListUploadsError>

Gets information about uploads, given an AWS Device Farm project ARN.

fn list_vpce_configurations(
&self,
input: ListVPCEConfigurationsRequest
) -> RusotoFuture<ListVPCEConfigurationsResult, ListVPCEConfigurationsError>

Returns information about all Amazon Virtual Private Cloud (VPC) endpoint configurations in the AWS account.

fn purchase_offering(
&self,
input: PurchaseOfferingRequest
) -> RusotoFuture<PurchaseOfferingResult, PurchaseOfferingError>

Immediately purchases offerings for an AWS account. Offerings renew with the latest total purchased quantity for an offering, unless the renewal was overridden. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

fn renew_offering(
&self,
input: RenewOfferingRequest
) -> RusotoFuture<RenewOfferingResult, RenewOfferingError>

Explicitly sets the quantity of devices to renew for an offering, starting from the effectiveDate of the next period. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

fn schedule_run(
&self,
input: ScheduleRunRequest
) -> RusotoFuture<ScheduleRunResult, ScheduleRunError>

Schedules a run.

fn stop_job(
&self,
input: StopJobRequest
) -> RusotoFuture<StopJobResult, StopJobError>

Initiates a stop request for the current job. AWS Device Farm will immediately stop the job on the device where tests have not started executing, and you will not be billed for this device. On the device where tests have started executing, Setup Suite and Teardown Suite tests will run to completion before stopping execution on the device. You will be billed for Setup, Teardown, and any tests that were in progress or already completed.

fn stop_remote_access_session(
&self,
input: StopRemoteAccessSessionRequest
) -> RusotoFuture<StopRemoteAccessSessionResult, StopRemoteAccessSessionError>

Ends a specified remote access session.

fn stop_run(
&self,
input: StopRunRequest
) -> RusotoFuture<StopRunResult, StopRunError>

Initiates a stop request for the current test run. AWS Device Farm will immediately stop the run on devices where tests have not started executing, and you will not be billed for these devices. On devices where tests have started executing, Setup Suite and Teardown Suite tests will run to completion before stopping execution on those devices. You will be billed for Setup, Teardown, and any tests that were in progress or already completed.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Associates the specified tags to a resource with the specified resourceArn. If existing tags on a resource are not specified in the request parameters, they are not changed. When a resource is deleted, the tags associated with that resource are deleted as well.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Deletes the specified tags from a resource.

fn update_device_instance(
&self,
input: UpdateDeviceInstanceRequest
) -> RusotoFuture<UpdateDeviceInstanceResult, UpdateDeviceInstanceError>

Updates information about an existing private device instance.

fn update_device_pool(
&self,
input: UpdateDevicePoolRequest
) -> RusotoFuture<UpdateDevicePoolResult, UpdateDevicePoolError>

Modifies the name, description, and rules in a device pool given the attributes and the pool ARN. Rule updates are all-or-nothing, meaning they can only be updated as a whole (or not at all).

fn update_instance_profile(
&self,
input: UpdateInstanceProfileRequest
) -> RusotoFuture<UpdateInstanceProfileResult, UpdateInstanceProfileError>

Updates information about an existing private device instance profile.

fn update_network_profile(
&self,
input: UpdateNetworkProfileRequest
) -> RusotoFuture<UpdateNetworkProfileResult, UpdateNetworkProfileError>

Updates the network profile with specific settings.

fn update_project(
&self,
input: UpdateProjectRequest
) -> RusotoFuture<UpdateProjectResult, UpdateProjectError>

Modifies the specified project name, given the project ARN and a new name.

fn update_upload(
&self,
input: UpdateUploadRequest
) -> RusotoFuture<UpdateUploadResult, UpdateUploadError>

Update an uploaded test specification (test spec).

fn update_vpce_configuration(
&self,
input: UpdateVPCEConfigurationRequest
) -> RusotoFuture<UpdateVPCEConfigurationResult, UpdateVPCEConfigurationError>

Updates information about an existing Amazon Virtual Private Cloud (VPC) endpoint configuration.

Loading content...