Crate rusoto_devicefarm

AWS Device Farm is a service that enables mobile app developers to test Android, iOS, and Fire OS apps on physical phones, tablets, and other devices in the cloud.

If you're using the service, you're probably looking for DeviceFarmClient and DeviceFarm.

Structs

AccountSettings
A container for account-level settings within AWS Device Farm.

Artifact
Represents the output of a test. Examples of artifacts include logs and screenshots.

CPU
Represents the amount of CPU that an app is using on a physical device.

Note that this does not represent system-wide CPU usage.

Counters
Represents entity counters.

CreateDevicePoolRequest
Represents a request to the create device pool operation.

CreateDevicePoolResult
Represents the result of a create device pool request.

CreateInstanceProfileRequest

CreateInstanceProfileResult

CreateNetworkProfileRequest

CreateNetworkProfileResult

CreateProjectRequest
Represents a request to the create project operation.

CreateProjectResult
Represents the result of a create project request.

CreateRemoteAccessSessionConfiguration
Configuration settings for a remote access session, including billing method.

CreateRemoteAccessSessionRequest
Creates and submits a request to start a remote access session.

CreateRemoteAccessSessionResult
Represents the server response from a request to create a remote access session.

CreateUploadRequest
Represents a request to the create upload operation.

CreateUploadResult
Represents the result of a create upload request.

CreateVPCEConfigurationRequest

CreateVPCEConfigurationResult

CustomerArtifactPaths
A JSON object specifying the paths where the artifacts generated by the customer's tests, on the device or in the test environment, will be pulled from.

Specify deviceHostPaths and optionally specify either iosPaths or androidPaths.

For web app tests, you can specify both iosPaths and androidPaths.

DeleteDevicePoolRequest
Represents a request to the delete device pool operation.

DeleteDevicePoolResult
Represents the result of a delete device pool request.

DeleteInstanceProfileRequest

DeleteInstanceProfileResult

DeleteNetworkProfileRequest

DeleteNetworkProfileResult

DeleteProjectRequest
Represents a request to the delete project operation.

DeleteProjectResult
Represents the result of a delete project request.

DeleteRemoteAccessSessionRequest
Represents the request to delete the specified remote access session.

DeleteRemoteAccessSessionResult
The response from the server when a request is made to delete the remote access session.

DeleteRunRequest
Represents a request to the delete run operation.

DeleteRunResult
Represents the result of a delete run request.

DeleteUploadRequest
Represents a request to the delete upload operation.

DeleteUploadResult
Represents the result of a delete upload request.

DeleteVPCEConfigurationRequest

DeleteVPCEConfigurationResult

Device
Represents a device type that an app is tested against.

DeviceFarmClient
A client for the AWS Device Farm API.

DeviceInstance
Represents the device instance.

DeviceMinutes
Represents the total (metered or unmetered) minutes used by the resource to run tests. Contains the sum of minutes consumed by all children.

DevicePool
Represents a collection of device types.

DevicePoolCompatibilityResult
Represents a device pool compatibility result.

ExecutionConfiguration
Represents configuration information about a test run, such as the execution timeout (in minutes).

GetAccountSettingsRequest
Represents the request sent to retrieve the account settings.

GetAccountSettingsResult
Represents the account settings return values from the GetAccountSettings request.

GetDeviceInstanceRequest

GetDeviceInstanceResult

GetDevicePoolCompatibilityRequest
Represents a request to the get device pool compatibility operation.

GetDevicePoolCompatibilityResult
Represents the result of describe device pool compatibility request.

GetDevicePoolRequest
Represents a request to the get device pool operation.

GetDevicePoolResult
Represents the result of a get device pool request.

GetDeviceRequest
Represents a request to the get device request.

GetDeviceResult
Represents the result of a get device request.

GetInstanceProfileRequest

GetInstanceProfileResult

GetJobRequest
Represents a request to the get job operation.

GetJobResult
Represents the result of a get job request.

GetNetworkProfileRequest

GetNetworkProfileResult

GetOfferingStatusRequest
Represents the request to retrieve the offering status for the specified customer or account.

GetOfferingStatusResult
Returns the status result for a device offering.

GetProjectRequest
Represents a request to the get project operation.

GetProjectResult
Represents the result of a get project request.

GetRemoteAccessSessionRequest
Represents the request to get information about the specified remote access session.

GetRemoteAccessSessionResult
Represents the response from the server that lists detailed information about the remote access session.

GetRunRequest
Represents a request to the get run operation.

GetRunResult
Represents the result of a get run request.

GetSuiteRequest
Represents a request to the get suite operation.

GetSuiteResult
Represents the result of a get suite request.

GetTestRequest
Represents a request to the get test operation.

GetTestResult
Represents the result of a get test request.

GetUploadRequest
Represents a request to the get upload operation.

GetUploadResult
Represents the result of a get upload request.

GetVPCEConfigurationRequest

GetVPCEConfigurationResult

IncompatibilityMessage
Represents information about incompatibility.

InstallToRemoteAccessSessionRequest
Represents the request to install an Android application (in .apk format) or an iOS application (in .ipa format) as part of a remote access session.

InstallToRemoteAccessSessionResult
Represents the response from the server after AWS Device Farm makes a request to install to a remote access session.

InstanceProfile
Represents the instance profile.

Job
Represents a device.

ListArtifactsRequest
Represents a request to the list artifacts operation.

ListArtifactsResult
Represents the result of a list artifacts operation.

ListDeviceInstancesRequest

ListDeviceInstancesResult

ListDevicePoolsRequest
Represents the result of a list device pools request.

ListDevicePoolsResult
Represents the result of a list device pools request.

ListDevicesRequest
Represents the result of a list devices request.

ListDevicesResult
Represents the result of a list devices operation.

ListInstanceProfilesRequest

ListInstanceProfilesResult

ListJobsRequest
Represents a request to the list jobs operation.

ListJobsResult
Represents the result of a list jobs request.

ListNetworkProfilesRequest

ListNetworkProfilesResult

ListOfferingPromotionsRequest

ListOfferingPromotionsResult

ListOfferingTransactionsRequest
Represents the request to list the offering transaction history.

ListOfferingTransactionsResult
Returns the transaction log of the specified offerings.

ListOfferingsRequest
Represents the request to list all offerings.

ListOfferingsResult
Represents the return values of the list of offerings.

ListProjectsRequest
Represents a request to the list projects operation.

ListProjectsResult
Represents the result of a list projects request.

ListRemoteAccessSessionsRequest
Represents the request to return information about the remote access session.

ListRemoteAccessSessionsResult
Represents the response from the server after AWS Device Farm makes a request to return information about the remote access session.

ListRunsRequest
Represents a request to the list runs operation.

ListRunsResult
Represents the result of a list runs request.

ListSamplesRequest
Represents a request to the list samples operation.

ListSamplesResult
Represents the result of a list samples request.

ListSuitesRequest
Represents a request to the list suites operation.

ListSuitesResult
Represents the result of a list suites request.

ListTestsRequest
Represents a request to the list tests operation.

ListTestsResult
Represents the result of a list tests request.

ListUniqueProblemsRequest
Represents a request to the list unique problems operation.

ListUniqueProblemsResult
Represents the result of a list unique problems request.

ListUploadsRequest
Represents a request to the list uploads operation.

ListUploadsResult
Represents the result of a list uploads request.

ListVPCEConfigurationsRequest

ListVPCEConfigurationsResult

Location
Represents a latitude and longitude pair, expressed in geographic coordinate system degrees (for example 47.6204, -122.3491).

Elevation is currently not supported.

MonetaryAmount
A number representing the monetary amount for an offering or transaction.

NetworkProfile
An array of settings that describes characteristics of a network profile.

Offering
Represents the metadata of a device offering.

OfferingPromotion
Represents information about an offering promotion.

OfferingStatus
The status of the offering.

OfferingTransaction
Represents the metadata of an offering transaction.

Problem
Represents a specific warning or failure.

ProblemDetail
Information about a problem detail.

Project
Represents an operating-system neutral workspace for running and managing tests.

PurchaseOfferingRequest
Represents a request for a purchase offering.

PurchaseOfferingResult
The result of the purchase offering (e.g., success or failure).

Radios
Represents the set of radios and their states on a device. Examples of radios include Wi-Fi, GPS, Bluetooth, and NFC.

RecurringCharge
Specifies whether charges for devices will be recurring.

RemoteAccessSession
Represents information about the remote access session.

RenewOfferingRequest
A request representing an offering renewal.

RenewOfferingResult
The result of a renewal offering.

Resolution
Represents the screen resolution of a device in height and width, expressed in pixels.

Rule
Represents a condition for a device pool.

Run
Represents a test run on a set of devices with a given app package, test parameters, etc.

Sample
Represents a sample of performance data.

ScheduleRunConfiguration
Represents the settings for a run. Includes things like location, radio states, auxiliary apps, and network profiles.

ScheduleRunRequest
Represents a request to the schedule run operation.

ScheduleRunResult
Represents the result of a schedule run request.

ScheduleRunTest
Represents additional test settings.

StopRemoteAccessSessionRequest
Represents the request to stop the remote access session.

StopRemoteAccessSessionResult
Represents the response from the server that describes the remote access session when AWS Device Farm stops the session.

StopRunRequest
Represents the request to stop a specific run.

StopRunResult
Represents the results of your stop run attempt.

Suite
Represents a collection of one or more tests.

Test
Represents a condition that is evaluated.

TrialMinutes
Represents information about free trial device minutes for an AWS account.

UniqueProblem
A collection of one or more problems, grouped by their result.

UpdateDeviceInstanceRequest

UpdateDeviceInstanceResult

UpdateDevicePoolRequest
Represents a request to the update device pool operation.

UpdateDevicePoolResult
Represents the result of an update device pool request.

UpdateInstanceProfileRequest

UpdateInstanceProfileResult

UpdateNetworkProfileRequest

UpdateNetworkProfileResult

UpdateProjectRequest
Represents a request to the update project operation.

UpdateProjectResult
Represents the result of an update project request.

UpdateVPCEConfigurationRequest

UpdateVPCEConfigurationResult

Upload
An app or a set of one or more tests to upload or that have been uploaded.

VPCEConfiguration
Represents an Amazon Virtual Private Cloud (VPC) endpoint configuration.

Enums

CreateDevicePoolError
Errors returned by CreateDevicePool

CreateInstanceProfileError
Errors returned by CreateInstanceProfile

CreateNetworkProfileError
Errors returned by CreateNetworkProfile

CreateProjectError
Errors returned by CreateProject

CreateRemoteAccessSessionError
Errors returned by CreateRemoteAccessSession

CreateUploadError
Errors returned by CreateUpload

CreateVPCEConfigurationError
Errors returned by CreateVPCEConfiguration

DeleteDevicePoolError
Errors returned by DeleteDevicePool

DeleteInstanceProfileError
Errors returned by DeleteInstanceProfile

DeleteNetworkProfileError
Errors returned by DeleteNetworkProfile

DeleteProjectError
Errors returned by DeleteProject

DeleteRemoteAccessSessionError
Errors returned by DeleteRemoteAccessSession

DeleteRunError
Errors returned by DeleteRun

DeleteUploadError
Errors returned by DeleteUpload

DeleteVPCEConfigurationError
Errors returned by DeleteVPCEConfiguration

GetAccountSettingsError
Errors returned by GetAccountSettings

GetDeviceError
Errors returned by GetDevice

GetDeviceInstanceError
Errors returned by GetDeviceInstance

GetDevicePoolCompatibilityError
Errors returned by GetDevicePoolCompatibility

GetDevicePoolError
Errors returned by GetDevicePool

GetInstanceProfileError
Errors returned by GetInstanceProfile

GetJobError
Errors returned by GetJob

GetNetworkProfileError
Errors returned by GetNetworkProfile

GetOfferingStatusError
Errors returned by GetOfferingStatus

GetProjectError
Errors returned by GetProject

GetRemoteAccessSessionError
Errors returned by GetRemoteAccessSession

GetRunError
Errors returned by GetRun

GetSuiteError
Errors returned by GetSuite

GetTestError
Errors returned by GetTest

GetUploadError
Errors returned by GetUpload

GetVPCEConfigurationError
Errors returned by GetVPCEConfiguration

InstallToRemoteAccessSessionError
Errors returned by InstallToRemoteAccessSession

ListArtifactsError
Errors returned by ListArtifacts

ListDeviceInstancesError
Errors returned by ListDeviceInstances

ListDevicePoolsError
Errors returned by ListDevicePools

ListDevicesError
Errors returned by ListDevices

ListInstanceProfilesError
Errors returned by ListInstanceProfiles

ListJobsError
Errors returned by ListJobs

ListNetworkProfilesError
Errors returned by ListNetworkProfiles

ListOfferingPromotionsError
Errors returned by ListOfferingPromotions

ListOfferingTransactionsError
Errors returned by ListOfferingTransactions

ListOfferingsError
Errors returned by ListOfferings

ListProjectsError
Errors returned by ListProjects

ListRemoteAccessSessionsError
Errors returned by ListRemoteAccessSessions

ListRunsError
Errors returned by ListRuns

ListSamplesError
Errors returned by ListSamples

ListSuitesError
Errors returned by ListSuites

ListTestsError
Errors returned by ListTests

ListUniqueProblemsError
Errors returned by ListUniqueProblems

ListUploadsError
Errors returned by ListUploads

ListVPCEConfigurationsError
Errors returned by ListVPCEConfigurations

PurchaseOfferingError
Errors returned by PurchaseOffering

RenewOfferingError
Errors returned by RenewOffering

ScheduleRunError
Errors returned by ScheduleRun

StopRemoteAccessSessionError
Errors returned by StopRemoteAccessSession

StopRunError
Errors returned by StopRun

UpdateDeviceInstanceError
Errors returned by UpdateDeviceInstance

UpdateDevicePoolError
Errors returned by UpdateDevicePool

UpdateInstanceProfileError
Errors returned by UpdateInstanceProfile

UpdateNetworkProfileError
Errors returned by UpdateNetworkProfile

UpdateProjectError
Errors returned by UpdateProject

UpdateVPCEConfigurationError
Errors returned by UpdateVPCEConfiguration

Traits

DeviceFarm
Trait representing the capabilities of the AWS Device Farm API. AWS Device Farm clients implement this trait.