Struct rusoto_devicefarm::DeviceFarmClient

pub struct DeviceFarmClient { /* fields omitted */ }

A client for the AWS Device Farm API.

Methods

impl DeviceFarmClient

pub fn new(region: Region) -> DeviceFarmClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> DeviceFarmClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,