Struct rusoto_devicefarm::ScheduleRunTest

pub struct ScheduleRunTest {
    pub filter: Option<String>,
    pub parameters: Option<HashMap<String, String>>,
    pub test_package_arn: Option<String>,
    pub test_spec_arn: Option<String>,
    pub type_: String,
}

Represents test settings. This data structure is passed in as the "test" parameter to ScheduleRun. For an example of the JSON request syntax, see ScheduleRun.

Fields
filter: Option<String>

The test's filter.

parameters: Option<HashMap<String, String>>

The test's parameters, such as test framework parameters and fixture settings. Parameters are represented by name-value pairs of strings.

For all tests:

appperformancemonitoring: Performance monitoring is enabled by default. Set this parameter to "false" to disable it.

For Calabash tests:

profile: A cucumber profile, for example, "myprofilename".

tags: You can limit execution to features or scenarios that have (or don't have) certain tags, for example, "@smoke" or "@smoke,~@wip".

For Appium tests (all types):

appiumversion: The Appium version. Currently supported values are "1.6.5" (and higher), "latest", and "default".

“latest” will run the latest Appium version supported by Device Farm (1.9.1).

For “default”, Device Farm will choose a compatible version of Appium for the device. The current behavior is to run 1.7.2 on Android devices and iOS 9 and earlier, 1.7.2 for iOS 10 and later.

This behavior is subject to change.

For Fuzz tests (Android only):

eventcount: The number of events, between 1 and 10000, that the UI fuzz test should perform.

throttle: The time, in ms, between 0 and 1000, that the UI fuzz test should wait between events.

seed: A seed to use for randomizing the UI fuzz test. Using the same seed value between tests ensures identical event sequences.

For Explorer tests:

username: A username to use if the Explorer encounters a login form. If not supplied, no username will be inserted.

password: A password to use if the Explorer encounters a login form. If not supplied, no password will be inserted.

For Instrumentation:

filter: A test filter string. Examples:

Running a single test case: "com.android.abc.Test1"

Running a single test: "com.android.abc.Test1#smoke"

Running multiple tests: "com.android.abc.Test1,com.android.abc.Test2"

For XCTest and XCTestUI:

filter: A test filter string. Examples:

Running a single test class: "LoginTests"

Running a multiple test classes: "LoginTests,SmokeTests"

Running a single test: "LoginTests/testValid"

Running multiple tests: "LoginTests/testValid,LoginTests/testInvalid"

For UIAutomator:

filter: A test filter string. Examples:

Running a single test case: "com.android.abc.Test1"

Running a single test: "com.android.abc.Test1#smoke"

Running multiple tests: "com.android.abc.Test1,com.android.abc.Test2"

test_package_arn: Option<String>

The ARN of the uploaded test that will be run.

test_spec_arn: Option<String>

The ARN of the YAML-formatted test specification.

type_: String

The test's type.

Must be one of the following values:

BUILTINFUZZ: The built-in fuzz type.

BUILTINEXPLORER: For Android, an app explorer that will traverse an Android app, interacting with it and capturing screenshots at the same time.

APPIUMJAVAJUNIT: The Appium Java JUnit type.

APPIUMJAVATESTNG: The Appium Java TestNG type.

APPIUMPYTHON: The Appium Python type.

APPIUMNODE: The Appium Node.js type.

APPIUMRUBY: The Appium Ruby type.

APPIUMWEBJAVAJUNIT: The Appium Java JUnit type for web apps.

APPIUMWEBJAVATESTNG: The Appium Java TestNG type for web apps.

APPIUMWEBPYTHON: The Appium Python type for web apps.

APPIUMWEBNODE: The Appium Node.js type for web apps.

APPIUMWEBRUBY: The Appium Ruby type for web apps.

CALABASH: The Calabash type.

INSTRUMENTATION: The Instrumentation type.

UIAUTOMATION: The uiautomation type.

UIAUTOMATOR: The uiautomator type.

XCTEST: The XCode test type.

XCTESTUI: The XCode UI test type.